move set_wd from tests/utils.py to utils.py
