+++ nf_core/components/components_test.py
@@ -19,7 +19,6 @@ from rich.text import Text
 
 import nf_core.utils
 from nf_core.components.components_command import ComponentCommand
-from tests.utils import set_wd
 
 log = logging.getLogger(__name__)
 
@@ -92,7 +91,7 @@ class ComponentsTest(ComponentCommand):  # type: ignore[misc]
         os.environ[
             "NFT_DIFF_ARGS"
         ] = "--line-numbers --expand-tabs=2"  # taken from https://code.askimed.com/nf-test/docs/assertions/snapshots/#snapshot-differences
-        with set_wd(Path(self.dir)):
+        with nf_core.utils.set_wd(Path(self.dir)):
             self.check_snapshot_stability()
         if len(self.errors) > 0:
             errors = "\n - ".join(self.errors)
